+++ lib/data/repositories/auth/auth_repository.dart
@@ -1,17 +1,7 @@
-import 'package:gogo_app/data/data_sources/auth/auth_data_source.dart';
-
 import 'package:gogo_app/data/models/auth/google_oauth/google_oauth_login_request.dart';
 import 'package:gogo_app/data/models/auth/google_oauth/google_oauth_login_response.dart';
-import 'auth_repository_impl.dart';
 
-class AuthRepositoryImpl implements AuthRepository {
-  final AuthDatasource _authDatasource;
-
-  AuthRepositoryImpl(this._authDatasource);
-
-  @override
+abstract class AuthRepository {
   Future<GoogleOAuthLoginResponse> googleOAuthLogin(
-      GoogleOAuthLoginRequest body) async {
-    return await _authDatasource.googleOAuthLogin(body);
-  }
+      GoogleOAuthLoginRequest body);
 }
