+++ lib/data/repositories/auth/auth_repository_impl.dart
@@ -1,6 +1,17 @@
 import 'package:gogo_app/data/models/auth/google_oauth/google_oauth_login_request.dart';
 import 'package:gogo_app/data/models/auth/google_oauth/google_oauth_login_response.dart';
 
-abstract class AuthRepository {
-  Future<GoogleOAuthLoginResponse> googleOAuthLogin(GoogleOAuthLoginRequest body);
+import '../../data_sources/auth/auth_data_source.dart';
+import 'auth_repository.dart';
+
+class AuthRepositoryImpl implements AuthRepository {
+  final AuthDatasource _authDatasource;
+
+  AuthRepositoryImpl(this._authDatasource);
+
+  @override
+  Future<GoogleOAuthLoginResponse> googleOAuthLogin(
+      GoogleOAuthLoginRequest body) async {
+    return await _authDatasource.googleOAuthLogin(body);
+  }
 }
